Enhance equality to address failing tests
Collections cannot easily be compared if underlying different implementations.
Consider: an UnmodifiableList is an UnmodifiableCollection BUT the same is not
          true in reverse.
The Mapbox Vector Tile Specification defines a list via the "repeat" protobuf
rule.

Vector tiles do not support Java int values but rather require longs.
Specifically, Objects.equals(1, 1L) == false.

A user attribute check has been added until an MvtAttributes class is defined.

+++ src/test/java/com/wdtinc/mapbox_vector_tile/adapt/jts/MvtEncoderTest.java
@@ -93,7 +93,9 @@ public final class MvtEncoderTest {
             Point point = GEOMETRY_FACTORY.createPoint(coord);
 
             Map<String, Object> attributes = new LinkedHashMap<>();
-            attributes.put("id", name.hashCode());
+            // Vector tiles only support 64 ints (Java long)
+            // https://github.com/mapbox/vector-tile-spec/blob/26b7823b16df1e15f55adb44243a3896acb7b804/2.1/vector_tile.proto#L24
+            attributes.put("id", (long) name.hashCode());
             attributes.put("name", name);
             point.setUserData(attributes);
